requirements.txt
click
websocket-client
termcolor
pytest
boto3
boto3-fixtures
moto
